sites/unesco/requirements/base.txt
boto3==1.28.9
django-check-seo==0.6.4
django-configurations==2.4.1
django-storages==1.13.2
dockerflow==2022.8.0
factory-boy==3.2.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
richie==2.24.1
sentry-sdk==1.28.1
mysqlclient==2.2.0
whitenoise==6.5.0
django-redis==5.3.0
